azure-function-schema-updater/requirements.txt
azure-functions
supabase
requests
